eliminated a findbugs warning about "int x int = long".
git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@11299 71c3de6d-444a-0410-be80-ed276b4c234a

+++ core/src/main/java/hudson/util/DoubleLaunchChecker.java
@@ -111,7 +111,7 @@ public class DoubleLaunchChecker {
      */
     public void schedule() {
         // randomize the scheduling so that multiple Hudson instances will write at the file at different time
-        int MINUTE = 1000*60;
+        long MINUTE = 1000*60;
         Trigger.timer.schedule(new SafeTimerTask() {
             protected void doRun() {
                 execute();
